Alpha 0.17
Added Debug Mode lights to use during building of the map.
Cleaned more of the project.
There is still 1.9% overlap of UV's on the hallway, still trying to find them.

I have named this Alpha 0.17 as its the 17th update. in the alpha build. This is the new version numbering system.