Doors!
Raycast doors, currently going to be working on locked doors.

Doors work as intended, just when you press E on the door, the light on the door turns green. On locked doors the light will flash red 3 times and give some form of indication to the user with both on screen (Door Locked) and error sounds.